test(sass): add address matcher
