+++ docs/source/tests/python/test/sass/instruction.rst
@@ -1,6 +1,7 @@
 Instruction
 ===========
 
+.. automodule:: tests.python.test.sass.instruction.test_address
 .. automodule:: tests.python.test.sass.instruction.test_fp32add
 .. automodule:: tests.python.test.sass.instruction.test_fp64add
 .. automodule:: tests.python.test.sass.instruction.test_register
