+++ reprospect/test/sass/composite_impl.py
@@ -9,8 +9,8 @@ import typing
 
 import mypy_extensions
 
-from reprospect.test.sass.instruction import InstructionMatcher, InstructionMatch, RegisterMatcher
-from reprospect.tools.sass            import Instruction
+from reprospect.test.sass.instruction import InstructionMatcher, InstructionMatch, AddressMatcher, RegisterMatcher
+from reprospect.tools.sass import Instruction
 
 if sys.version_info >= (3, 12):
     from typing import override
@@ -225,7 +225,7 @@ class AnyOfMatcher(SequenceMatcher):
     def explain(self, *, instructions : typing.Sequence[Instruction | str]) -> str:
         return f'None of {self.matchers!r} did match {instructions}.'
 
-OperandMatcher : typing.TypeAlias = str | RegisterMatcher
+OperandMatcher : typing.TypeAlias = str | AddressMatcher | RegisterMatcher
 
 class OperandValidator(InstructionMatcher):
     """
@@ -252,7 +252,7 @@ class OperandValidator(InstructionMatcher):
                 return None
             if isinstance(self.operand, str) and operand == self.operand:
                 return matched
-            if isinstance(self.operand, RegisterMatcher) and self.operand.match(operand) is not None:
+            if isinstance(self.operand, AddressMatcher | RegisterMatcher) and self.operand.match(operand) is not None:
                 return matched
         return None
 
@@ -281,7 +281,7 @@ class OperandsValidator(InstructionMatcher):
                     return None
                 if isinstance(moperand, str) and operand != moperand:
                     return None
-                if isinstance(moperand, RegisterMatcher) and moperand.match(operand) is None:
+                if isinstance(moperand, AddressMatcher | RegisterMatcher) and moperand.match(operand) is None:
                     return None
             return matched
         return None
