+++ tests/python/test/sass/instruction/CMakeLists.txt
@@ -1,3 +1,4 @@
+add_pytest(address)
 add_pytest(fp32add)
 add_pytest(fp64add)
 add_pytest(register)
