+++ tests/python/test/sass/test_instruction.py
@@ -96,117 +96,6 @@ class TestPatternBuilder:
         assert regex.match(PatternBuilder.UREG,   'UR42') is not None
         assert regex.match(PatternBuilder.ureg(), 'UR42') is not None
 
-    def test_address_reg(self):
-        """
-        Test :py:meth:`reprospect.test.sass.instruction.PatternBuilder.address` for :py:attr:`reprospect.test.sass.instruction.PatternBuilder.REG`.
-        """
-        # Defaulted offset behaviour.
-        pattern = PatternBuilder.address(PatternBuilder.REG)
-
-        matched = regex.match(pattern, string = '[R0]')
-        assert matched is not None
-        assert matched.group('address')  == 'R0' and matched.captures('address')  == ['R0']
-        assert matched.group('operands') == 'R0' and matched.captures('operands') == ['R0']
-
-        matched = regex.match(pattern, string = '[R2+0x10]')
-        assert matched is not None
-        assert matched.group('address')  == 'R2+0x10' and matched.captures('address')  == ['R2+0x10']
-        assert matched.group('operands') == 'R2+0x10' and matched.captures('operands') == ['R2+0x10']
-
-        # Enforced offset behaviour.
-        pattern = PatternBuilder.address(PatternBuilder.REG, offset = True)
-
-        assert regex.match(pattern, string = '[R0]') is None
-
-        matched = regex.match(pattern, string = '[R2+0x10]')
-        assert matched is not None
-        assert matched.group('address')  == 'R2+0x10' and matched.captures('address')  == ['R2+0x10']
-        assert matched.group('operands') == 'R2+0x10' and matched.captures('operands') == ['R2+0x10']
-
-        # Disabled offset behaviour.
-        pattern = PatternBuilder.address(PatternBuilder.REG, offset = False)
-
-        matched = regex.match(pattern, string = '[R0]')
-        assert matched is not None
-        assert matched.group('address')  == 'R0' and matched.captures('address')  == ['R0']
-        assert matched.group('operands') == 'R0' and matched.captures('operands') == ['R0']
-
-        assert regex.match(pattern, string = '[R2+0x10]') is None
-
-    def test_address_reg64(self):
-        """
-        Test :py:meth:`reprospect.test.sass.instruction.PatternBuilder.address` for :py:attr:`reprospect.test.sass.instruction.PatternBuilder.REG64`.
-        """
-        # Default offset behaviour.
-        pattern = PatternBuilder.address(PatternBuilder.REG64)
-
-        matched = regex.match(pattern, string = '[R0.64]')
-        assert matched is not None
-        assert matched.group('address')  == 'R0.64' and matched.captures('address')  == ['R0.64']
-        assert matched.group('operands') == 'R0.64' and matched.captures('operands') == ['R0.64']
-
-        matched = regex.match(pattern, string = '[R2.64+0x10]')
-        assert matched is not None
-        assert matched.group('address')  == 'R2.64+0x10' and matched.captures('address')  == ['R2.64+0x10']
-        assert matched.group('operands') == 'R2.64+0x10' and matched.captures('operands') == ['R2.64+0x10']
-
-        # Enforced offset behaviour.
-        pattern = PatternBuilder.address(PatternBuilder.REG64, offset = True)
-
-        assert regex.match(pattern, string = '[R0.64]') is None
-
-        matched = regex.match(pattern, string = '[R2.64+0x10]')
-        assert matched is not None
-        assert matched.group('address')  == 'R2.64+0x10' and matched.captures('address')  == ['R2.64+0x10']
-        assert matched.group('operands') == 'R2.64+0x10' and matched.captures('operands') == ['R2.64+0x10']
-
-        # Disabled offset behaviour.
-        pattern = PatternBuilder.address(PatternBuilder.REG64, offset = False)
-
-        matched = regex.match(pattern, string = '[R0.64]')
-        assert matched is not None
-        assert matched.group('address')  == 'R0.64' and matched.captures('address')  == ['R0.64']
-        assert matched.group('operands') == 'R0.64' and matched.captures('operands') == ['R0.64']
-
-        assert regex.match(pattern, string = '[R2.64+0x10]') is None
-
-    def test_desc_reg64_addr(self):
-        """
-        Test :py:meth`reprospect.test.sass.instruction.PatternBuilder.desc_reg64_addr`.
-        """
-        # Default offset behaviour.
-        pattern = PatternBuilder.desc_reg64_addr()
-
-        matched = regex.match(pattern, string = 'desc[UR86][R0.64]')
-        assert matched is not None
-        assert matched.group('address')  == 'desc[UR86][R0.64]' and matched.captures('address')  == ['desc[UR86][R0.64]']
-        assert matched.group('operands') == 'desc[UR86][R0.64]' and matched.captures('operands') == ['desc[UR86][R0.64]']
-
-        matched = regex.match(pattern, string = 'desc[UR86][R2.64+0x10]')
-        assert matched is not None
-        assert matched.group('address')  == 'desc[UR86][R2.64+0x10]' and matched.captures('address')  == ['desc[UR86][R2.64+0x10]']
-        assert matched.group('operands') == 'desc[UR86][R2.64+0x10]' and matched.captures('operands') == ['desc[UR86][R2.64+0x10]']
-
-        # Enforced offset behaviour.
-        pattern = PatternBuilder.desc_reg64_addr(offset = True)
-
-        assert regex.match(pattern, string = 'desc[UR86][R0.64]') is None
-
-        matched = regex.match(pattern, string = 'desc[UR86][R2.64+0x10]')
-        assert matched is not None
-        assert matched.group('address')  == 'desc[UR86][R2.64+0x10]' and matched.captures('address')  == ['desc[UR86][R2.64+0x10]']
-        assert matched.group('operands') == 'desc[UR86][R2.64+0x10]' and matched.captures('operands') == ['desc[UR86][R2.64+0x10]']
-
-        # Disabled offset behaviour.
-        pattern = PatternBuilder.desc_reg64_addr(offset = False)
-
-        matched = regex.match(pattern, string = 'desc[UR86][R0.64]')
-        assert matched is not None
-        assert matched.group('address')  == 'desc[UR86][R0.64]' and matched.captures('address')  == ['desc[UR86][R0.64]']
-        assert matched.group('operands') == 'desc[UR86][R0.64]' and matched.captures('operands') == ['desc[UR86][R0.64]']
-
-        assert regex.match(pattern, string = 'desc[UR86][R2.64+0x10]') is None
-
     def test_anygpreg(self) -> None:
         """
         Test :py:meth:`reprospect.test.sass.instruction.PatternBuilder.anygpreg`.
